Added tests, renamed pagenumber to page-number
